pyproject.toml
[tool.poetry]
name = "glimpse-portal"
version = "0.1.0"
description = "Django admin portal for Glimpse App - connected to Supabase"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "~5.1"
gunicorn = "^23.0"
psycopg = {extras = ["binary"], version = "^3.2"}
dj-database-url = "^2.3"
python-decouple = "^3.8"
whitenoise = "^6.8"
django-cors-headers = "^4.6"
djangorestframework = "^3.15"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
